refine graphs submodules scope
