+++ tests/test_graphs/test_base_graph.py
@@ -154,6 +154,136 @@ def test_basegraph_constructors(n_nodes: int) -> None:
     assert graph2.is_ud_graph()
 
 
+@pytest.mark.parametrize("n_nodes", [5, 10, 50])
+def test_from_matrix(n_nodes: int) -> None:
+    np.random.seed(0)
+    data = np.random.rand(n_nodes, n_nodes)
+
+    with pytest.raises(ValueError):
+        # Matrix is not symmetric
+        graph = BaseGraph.from_matrix(data)
+
+    data = data + data.T
+    data_copy = data.copy()  # Make a copy to test the original data
+
+    graph = BaseGraph.from_matrix(data)
+
+    # Check input data matrix has not been modified
+    np.testing.assert_equal(data, data_copy)
+
+    assert len(graph.node_weights) == graph.number_of_nodes()
+    assert len(graph.edge_weights) == graph.number_of_edges()
+    assert graph.has_node_weights
+    assert graph.has_edge_weights
+
+    data_diag = np.diag(data)
+    node_weights = list(graph.node_weights.values())
+    np.testing.assert_allclose(node_weights, data_diag)
+
+    # Build a fresh matrix for the second sub-test: zero out the diagonal and some random edges
+    almost_zero = 1e-14
+    data2 = data_copy.copy()
+    np.fill_diagonal(data2, almost_zero)
+    random_edges_removal = random_edge_list(range(n_nodes), k=4)
+    i_list, j_list = zip(*random_edges_removal)
+    data2[i_list, j_list] = almost_zero
+    data2[j_list, i_list] = almost_zero
+
+    graph = BaseGraph.from_matrix(data2)
+
+    assert not graph.has_node_weights
+    assert graph.has_edge_weights
+
+    for edge in random_edges_removal:
+        assert edge not in graph.sorted_edges
+
+    n_edges = graph.number_of_edges()
+    idx = [2 * i for i in range(n_edges)]
+
+    data_edge_weights = np.sort(data2[data2 >= 1e-7])[idx]
+    edge_weights = sorted(list(graph.edge_weights.values()))
+
+    np.testing.assert_allclose(edge_weights, data_edge_weights)
+
+
+@pytest.mark.parametrize("n_nodes", [5, 10, 50])
+def test_to_matrix_unweighted(n_nodes: int) -> None:
+    graph = BaseGraph.from_nodes(range(n_nodes))
+    graph.add_edges_from(random_edge_list(range(n_nodes), k=2 * n_nodes))
+    # FIXME: _edge_weights is a snapshot that goes stale after add_edges_from;
+    # see issue #431 (edge_weights does not reflect edges added after construction).
+    graph._reset_dicts()
+    assert not graph.has_node_weights
+    assert not graph.has_edge_weights
+
+    matrix = graph.to_matrix()
+
+    np.testing.assert_equal(matrix, matrix.T)
+    np.testing.assert_equal(np.diag(matrix), np.zeros(n_nodes))
+
+    for i, j in graph.sorted_edges:
+        assert matrix[i, j] == 1.0
+        assert matrix[j, i] == 1.0
+
+    non_edges = graph.all_node_pairs - graph.sorted_edges
+    for i, j in non_edges:
+        assert matrix[i, j] == 0.0
+        assert matrix[j, i] == 0.0
+
+
+@pytest.mark.parametrize("n_nodes", [5, 10, 50])
+def test_to_matrix_weighted(n_nodes: int) -> None:
+    graph = BaseGraph.from_nodes(range(n_nodes))
+    graph.add_edges_from(random_edge_list(range(n_nodes), k=2 * n_nodes))
+    graph.node_weights = {i: np.random.rand() for i in graph.nodes}
+    graph.edge_weights = {e: np.random.rand() for e in graph.sorted_edges}
+
+    matrix = graph.to_matrix()
+
+    np.testing.assert_equal(matrix, matrix.T)
+    np.testing.assert_allclose(np.diag(matrix), list(graph.node_weights.values()))
+
+    for (i, j), weight in graph.edge_weights.items():
+        assert matrix[i, j] == weight
+        assert matrix[j, i] == weight
+
+
+@pytest.mark.parametrize("n_nodes", [3, 7, 21])
+def test_to_matrix_roundtrip(n_nodes: int) -> None:
+    rng = np.random.default_rng(12345)
+    matrix = rng.normal(0, 1, size=(n_nodes, n_nodes))
+    # zero out some elements for testing
+    rows, cols = rng.integers(n_nodes, size=3), rng.integers(n_nodes, size=3)
+    matrix[rows, cols] = 0.0
+    matrix[cols, rows] = 0.0
+    matrix += matrix.T
+
+    graph = BaseGraph.from_matrix(matrix)
+    np.testing.assert_allclose(graph.to_matrix(), matrix, atol=1e-8)
+
+
+def test_to_matrix_custom_node_labels_and_none_weights() -> None:
+    # Ensure `to_matrix()` respects `self.nodes` ordering and handles None weights.
+    # Also ensures it works with non-0..N-1 node labels.
+    graph = BaseGraph.from_nodes(["b", "a", "c"])  # preserve explicit order
+    graph.add_edges_from([("b", "a"), ("a", "c")])
+
+    # Mix of real weights and None
+    graph.node_weights = {"b": 2.0, "a": None, "c": -3.0}
+    # NOTE: order is not guaranteed because BaseGraph maintains arbitrarily sorted edges
+    graph.edge_weights = {("a", "b"): None, ("a", "c"): 0.25}
+
+    matrix = graph.to_matrix()
+    expected = np.array(
+        [
+            [2.0, 1.0, 0.0],
+            [1.0, 0.0, 0.25],
+            [0.0, 0.25, -3.0],
+        ],
+    )
+    np.testing.assert_equal(matrix, expected)
+
+
 @pytest.mark.parametrize("input", ["hello", Data()])
 def test_from_nx_wrong_input(input: Any) -> None:
     with pytest.raises(TypeError, match="Input must be a networkx.Graph instance."):
