+++ tests/test_graphs/test_data_graph.py
@@ -6,9 +6,7 @@ import pytest
 import torch
 from torch_geometric.data import Data
 
-from qoolqit.graphs import DataGraph, random_edge_list
-
-ATOL_64 = 1e-14
+from qoolqit.graphs import DataGraph
 
 
 @pytest.mark.parametrize("graph_type", ["circle", "line", "random_ud"])
@@ -65,131 +63,6 @@ def test_datagraph_random_er(n_nodes: int) -> None:
         graph.distances()
     with pytest.raises(AttributeError):
         graph.min_distance()
-
-
-@pytest.mark.parametrize("n_nodes", [5, 10, 50])
-def test_datagraph_from_matrix(n_nodes: int) -> None:
-    np.random.seed(0)
-    data = np.random.rand(n_nodes, n_nodes)
-
-    with pytest.raises(ValueError):
-        # Matrix is not symmetric
-        graph = DataGraph.from_matrix(data)
-
-    data = data + data.T
-    data_copy = data.copy()  # Make a copy to test the original data
-
-    graph = DataGraph.from_matrix(data)
-
-    # Check input data matrix has not been modified
-    np.testing.assert_equal(data, data_copy)
-
-    assert len(graph.node_weights) == graph.number_of_nodes()
-    assert len(graph.edge_weights) == graph.number_of_edges()
-    assert graph.has_node_weights
-    assert graph.has_edge_weights
-
-    data_diag = np.diag(data)
-    node_weights = list(graph.node_weights.values())
-    np.testing.assert_allclose(node_weights, data_diag)
-
-    # Build a fresh matrix for the second sub-test: zero out the diagonal and some random edges
-    almost_zero = ATOL_64
-    data2 = data_copy.copy()
-    np.fill_diagonal(data2, almost_zero)
-    random_edges_removal = random_edge_list(range(n_nodes), k=4)
-    i_list, j_list = zip(*random_edges_removal)
-    data2[i_list, j_list] = almost_zero
-    data2[j_list, i_list] = almost_zero
-
-    graph = DataGraph.from_matrix(data2)
-
-    assert not graph.has_node_weights
-    assert graph.has_edge_weights
-
-    for edge in random_edges_removal:
-        assert edge not in graph.sorted_edges
-
-    n_edges = graph.number_of_edges()
-    idx = [2 * i for i in range(n_edges)]
-
-    data_edge_weights = np.sort(data2[data2 >= 1e-7])[idx]
-    edge_weights = sorted(list(graph.edge_weights.values()))
-
-    np.testing.assert_allclose(edge_weights, data_edge_weights)
-
-
-@pytest.mark.parametrize("n_nodes", [5, 10, 50])
-def test_datagraph_to_matrix_unweighted(n_nodes: int) -> None:
-    graph = DataGraph.random_er(n_nodes, p=0.5, seed=0)
-    assert not graph.has_node_weights
-    assert not graph.has_edge_weights
-
-    matrix = graph.to_matrix()
-
-    np.testing.assert_equal(matrix, matrix.T)
-    np.testing.assert_equal(np.diag(matrix), np.zeros(n_nodes))
-
-    for i, j in graph.sorted_edges:
-        assert matrix[i, j] == 1.0
-        assert matrix[j, i] == 1.0
-
-    non_edges = graph.all_node_pairs - graph.sorted_edges
-    for i, j in non_edges:
-        assert matrix[i, j] == 0.0
-        assert matrix[j, i] == 0.0
-
-
-@pytest.mark.parametrize("n_nodes", [5, 10, 50])
-def test_datagraph_to_matrix_weighted(n_nodes: int) -> None:
-    graph = DataGraph.random_er(n_nodes, p=0.5, seed=0)
-    graph.node_weights = {i: np.random.rand() for i in graph.nodes}
-    graph.edge_weights = {e: np.random.rand() for e in graph.sorted_edges}
-
-    matrix = graph.to_matrix()
-
-    np.testing.assert_equal(matrix, matrix.T)
-    np.testing.assert_allclose(np.diag(matrix), list(graph.node_weights.values()))
-
-    for (i, j), weight in graph.edge_weights.items():
-        assert matrix[i, j] == weight
-        assert matrix[j, i] == weight
-
-
-@pytest.mark.parametrize("n_nodes", [3, 7, 21])
-def test_datagraph_to_matrix_roundtrip(n_nodes: int) -> None:
-    rng = np.random.default_rng(12345)
-    matrix = rng.normal(0, 1, size=(n_nodes, n_nodes))
-    # zero out some elements for testing
-    rows, cols = rng.integers(n_nodes, size=3), rng.integers(n_nodes, size=3)
-    matrix[rows, cols] = 0.0
-    matrix[cols, rows] = 0.0
-    matrix += matrix.T
-
-    graph = DataGraph.from_matrix(matrix)
-    np.testing.assert_allclose(graph.to_matrix(), matrix, atol=1e-8)
-
-
-def test_datagraph_to_matrix_custom_node_labels_and_none_weights() -> None:
-    # Ensure `to_matrix()` respects `self.nodes` ordering and handles None weights.
-    # Also ensures it works with non-0..N-1 node labels.
-    graph = DataGraph.from_nodes(["b", "a", "c"])  # preserve explicit order
-    graph.add_edges_from([("b", "a"), ("a", "c")])
-
-    # Mix of real weights and None
-    graph.node_weights = {"b": 2.0, "a": None, "c": -3.0}
-    # TOFIX: order is not guaranteed because DataGraph maintains arbitrarily sort edges
-    graph.edge_weights = {("a", "b"): None, ("a", "c"): 0.25}
-
-    matrix = graph.to_matrix()
-    expected = np.array(
-        [
-            [2.0, 1.0, 0.0],
-            [1.0, 0.0, 0.25],
-            [0.0, 0.25, -3.0],
-        ],
-    )
-    np.testing.assert_equal(matrix, expected)
 
 
 def test_triangular() -> None:
